[m2n] add AFDExtraFields for different afdconnectors
